requirements.txt
pandas
dotenv
pymongo[srv]
scikit-learn
numpy
mlflow
dagshub
fastapi
Flask
uvicorn
dataclasses
python-box